Refactor README extraction into it's own class
